chore: create `Reminder` subfolder & move `Dtos` and `ValueObjects` folder
