+++ Backend/src/BuildingBlocks/BuildingBlocks.Application/Data/INodesService.cs
@@ -1,6 +1,6 @@
 using BuildingBlocks.Domain.Nodes.Node.Dtos;
 using BuildingBlocks.Domain.Nodes.Node.ValueObjects;
-using BuildingBlocks.Domain.Reminders.ValueObjects;
+using BuildingBlocks.Domain.Reminders.Reminder.ValueObjects;
 
 namespace BuildingBlocks.Application.Data;
