+++ Backend/src/BuildingBlocks/BuildingBlocks.Application/Data/IRemindersService.cs
@@ -1,5 +1,5 @@
-using BuildingBlocks.Domain.Reminders.Dtos;
-using BuildingBlocks.Domain.Reminders.ValueObjects;
+using BuildingBlocks.Domain.Reminders.Reminder.Dtos;
+using BuildingBlocks.Domain.Reminders.Reminder.ValueObjects;
 
 namespace BuildingBlocks.Application.Data;
