+++ Backend/src/BuildingBlocks/BuildingBlocks.Domain/Reminders/Reminder/Dtos/ReminderDto.cs
@@ -1,15 +1,17 @@
 ﻿using System.Text.Json.Serialization;
+using BuildingBlocks.Domain.Nodes.Node.Dtos;
 
-namespace BuildingBlocks.Domain.Reminders.Dtos;
+namespace BuildingBlocks.Domain.Reminders.Reminder.Dtos;
 
-public class ReminderBaseDto(
+public class ReminderDto(
     string? id,
     string title,
     string description,
     DateTime dueDateTime,
     int priority,
     DateTime notificationTime,
-    string status)
+    string status,
+    NodeBaseDto node) // todo: Refactor this so that it extends ReminderBaseDto
 {
     [JsonPropertyName("id")] public string? Id { get; } = id;
 
@@ -24,4 +26,6 @@ public class ReminderBaseDto(
     [JsonPropertyName("notificationTime")] public DateTime NotificationTime { get; } = notificationTime;
 
     [JsonPropertyName("status")] public string Status { get; } = status;
+
+    [JsonPropertyName("node")] public NodeBaseDto Node { get; set; } = node;
 }
