+++ Backend/src/Modules/Reminders/Reminders.Api/Endpoints/Reminders/CreateReminder.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using BuildingBlocks.Domain.Nodes.Node.ValueObjects;
-using BuildingBlocks.Domain.Reminders.ValueObjects;
+using BuildingBlocks.Domain.Reminders.Reminder.ValueObjects;
 using Reminders.Application.Entities.Reminders.Commands.CreateReminder;
 
 // ReSharper disable ClassNeverInstantiated.Global
