+++ Backend/src/Modules/Reminders/Reminders.Api/Endpoints/Reminders/ListReminders.cs
@@ -1,5 +1,5 @@
 ﻿using BuildingBlocks.Application.Pagination;
-using BuildingBlocks.Domain.Reminders.Dtos;
+using BuildingBlocks.Domain.Reminders.Reminder.Dtos;
 using Reminders.Application.Entities.Reminders.Queries.ListReminders;
 
 namespace Reminders.Api.Endpoints.Reminders;
