+++ Backend/src/Modules/Reminders/Reminders.Application/Data/Abstractions/IRemindersRepository.cs
@@ -1,4 +1,4 @@
-using BuildingBlocks.Domain.Reminders.ValueObjects;
+using BuildingBlocks.Domain.Reminders.Reminder.ValueObjects;
 
 namespace Reminders.Application.Data.Abstractions;
