+++ Backend/src/Modules/Reminders/Reminders.Application/Entities/Reminders/Commands/CreateReminder/CreateReminderCommand.cs
@@ -1,5 +1,5 @@
 ﻿using BuildingBlocks.Domain.Nodes.Node.ValueObjects;
-using BuildingBlocks.Domain.Reminders.ValueObjects;
+using BuildingBlocks.Domain.Reminders.Reminder.ValueObjects;
 
 namespace Reminders.Application.Entities.Reminders.Commands.CreateReminder;
