+++ Backend/src/Modules/Reminders/Reminders.Application/Entities/Reminders/Commands/DeleteReminder/DeleteReminderCommand.cs
@@ -1,4 +1,4 @@
-﻿using BuildingBlocks.Domain.Reminders.ValueObjects;
+﻿using BuildingBlocks.Domain.Reminders.Reminder.ValueObjects;
 
 namespace Reminders.Application.Entities.Reminders.Commands.DeleteReminder;
