+++ Backend/src/Modules/Reminders/Reminders.Infrastructure/Data/Configurations/ReminderConfiguration.cs
@@ -1,4 +1,4 @@
-﻿using BuildingBlocks.Domain.Reminders.ValueObjects;
+﻿using BuildingBlocks.Domain.Reminders.Reminder.ValueObjects;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
 
 namespace Reminders.Infrastructure.Data.Configurations;
