+++ Backend/src/Modules/Reminders/Reminders.Infrastructure/Data/Extensions/InitialData.cs
@@ -1,5 +1,5 @@
 ﻿using BuildingBlocks.Domain.Nodes.Node.ValueObjects;
-using BuildingBlocks.Domain.Reminders.ValueObjects;
+using BuildingBlocks.Domain.Reminders.Reminder.ValueObjects;
 
 namespace Reminders.Infrastructure.Data.Extensions;
